NO-TICKET: execution-engine/engine-core: restructure modules
