+++ execution-engine/engine-core/src/tracking_copy/byte_size.rs
