+++ execution-engine/engine-core/src/tracking_copy/meter.rs
@@ -4,7 +4,8 @@ pub trait Meter<K, V> {
 }
 
 pub mod heap_meter {
-    use crate::byte_size::ByteSize;
+    use tracking_copy::byte_size::ByteSize;
+
     pub struct HeapSize;
 
     impl<K: ByteSize, V: ByteSize> super::Meter<K, V> for HeapSize {
