+++ execution-engine/engine-grpc-server/src/engine_server/mappings/mod.rs
@@ -15,7 +15,7 @@ use engine_core::engine_state::execution_effect::ExecutionEffect;
 use engine_core::engine_state::execution_result::ExecutionResult;
 use engine_core::engine_state::op::Op;
 use engine_core::execution::Error as ExecutionError;
-use engine_core::utils;
+use engine_core::tracking_copy::utils;
 use engine_server::{ipc, state};
 use engine_shared::logging;
 use engine_shared::logging::log_level;
